fix warnings

+++ Project1/FragmentShader.cpp
@@ -1,5 +1,6 @@
 #include "FragmentShader.h"
 #include <cmath>
+#include <algorithm>
 
 Vector4 FragmentShader::ShaderProc(VertexOut v, const ShadingConstants& constants)
 {
@@ -7,10 +8,10 @@ Vector4 FragmentShader::ShaderProc(VertexOut v, const ShadingConstants& constant
     Vector4 view = (constants.cameraPos - v.worldPos).Normalize();
     Vector4 h = (view + constants.lightDir).Normalize();
 
-    float ambientStrength = 0.1;
+    float ambientStrength = 0.1f;
     v.worldNormal = v.worldNormal.Normalize();
     float diff = std::max(0.0f, v.worldNormal * constants.lightDir);
-    float spec = pow(std::max(0.0f, v.worldNormal * h), 50);
+    float spec = (float)pow(std::max(0.0f, v.worldNormal * h), 50);
     Vector4 diffuse = constants.lightColor * diff * constants.intensity;
     Vector4 ambient = constants.lightColor * ambientStrength;
     Vector4 specular = constants.lightColor * spec;
