+++ Project1/Matrix.cpp
@@ -31,12 +31,12 @@ Matrix Matrix::Scale(float x, float y, float z)
 
 Matrix Matrix::Rotate(float degree, float x, float y, float z)
 {
-	float radian = degree * PI / 180.0f;
+	double radian = degree * PI / 180.0f;
 	Vector4 v{ x, y, z, 0.0f };
 	Vector4 unitV = v.Normalize();
-	float s = sin(radian);
-	float c = cos(radian);
-	Vector4 vcos = unitV * (1.0 - c);
+	float s = (float)sin(radian);
+	float c = (float)cos(radian);
+	Vector4 vcos = unitV * (1.0f - c);
 
 
 	Matrix  result({ vcos.x * unitV.x + c, vcos.x * unitV.y - s * unitV.z, vcos.x * unitV.z + s * unitV.y, 0.0f },
